Add a new hash class in __gnu_ext for the extension containers. There are two motivations for this.
First, this allows users who are specializing __gnu_ext::hash to
continue doing so without changing their code.

Second, SGI specifies hash overloads for char* and const char* that
perform a hash of the string, not of the pointer.

In order to support this, the hashing code for string is factored out.

llvm-svn: 136539
GitOrigin-RevId: 8d2ed56644047c08e2646aa8c981bebd84347b4a