win32: temporarily disable setting locale on 14+
The locale structures have been made opaque in CRT 14+.  This currently
prevents building libc++ for Windows.  We can re-enable this in the
future when we have replicated the structure to access the private field
for the name (unless there exists a better supported mechanism to query
the name of a locale given the locale_t).

llvm-svn: 290835
GitOrigin-RevId: 11c0c53367efbcc8d54ff26c562e8b7891ebc1b1