Don't require relops on variant alternatives to all return the same type.
Libc++ correctly asserts that a set of visitors for a variant all
return the same type. However, we use the visitation machinary to
perform relational operations. This causes a static assertion when
some of the alternatives relops return a UDT which is implicitly
convertible to bool instead of 'bool' exactly.

llvm-svn: 342560
GitOrigin-RevId: 7dca3127c2ad98c3f5a291d8f4878348772bb558